Cleaned Code
Simplified the process of checking if InfiniBot has permissions to send messages inside a channel, and then warning the user if not. Now, there is a single function that handles all of that.

A lot of the other changes were just me implementing the function.

Also moved some functions for cleaner code.