Non-active reaction roles now actually work.
Changed the algorithm that determines whether or not a message is a reaction role. Previously, if it was not an active message any more, it was no longer considered a reaction role. This has been changed to instead look at the first field and if it starts with the word "react".