Intermediary commit message
Working on minor things here and there, including
spelling issues and fixing some bugs.

Currently working on storing active messages that
the bot sends so that it can edit them in the
future.

THIS COMMIT IS ONLY MADE FOR THE PURPOSE OF TAKING
THIS PROJECT OUT OF A CODESPACE. THIS COMMIT DOES
NOT REPRESENT A COMPLETE CHANGE.