Updated tests
Improved testing code to ensure that tests are better isolated. In
addition, I added some tests for Stored Messages, and that actually
revealed some issues that I fixed.

I renamed created_at to last_updated because I felt that if a message is
edited, that timestamp should be refreshed so that it's active for
longer. I thought that last_updated was more representative of that
purposee.

I also added some comparison functions to the StoredMessages class so
that lists of them can be compared (useful for the tests)

I notice that I'm only testing half of the cleanup function, so I'll
work on that now.