Fixed Theme Change Button in Pages Site
Previously, the theme change button was coded to not take up any
vertical space when a page lacked breadcrumbs. However, that caused some
formatting issues on mobile, so that change has been reverted.

The theme change button will always take up some amount of space on the
top of the screen now, and formatting issues are solved.