Updated Profane Words
I updated the default profane words to consist of accurate, real,
profane and NSFW words. The thought process behind this was that if
someone wants to clone the bot and use it themselves, their default
filtered words will already be setup.

I also had to update the dashboard to use a codeblock so that the
special symbols appeared properly.