Closed Strange Bug when Updating Channels
This bug was introduced in f34d4e5, and it centered around me removing
the __dict__ definition for SimpleTableManager(), and renaming it to
to_dict(). I thought all dependencies for that were removed, but it
directly led to a failure to send a message to the updated Admin/Log
channel when updating the channel in the db, since it somehow caused the
cached value type to not be implicitly converted? Still not completely
sure what happened, but this change fixes it.

I also modernized the code of the Admin Channel Setter page of the
dashboard.