Closed Bug with Moderation Daily Maintenance
There was a weird bug where the format for the strike_info.last_strike
wasn't correct, and it was causing an error in one server. I'm unsure
how it got to be in that state, but this change allows the code to
process and fix it.

ERROR when checking profanity moderation (member) in server
[redacted-server-id]: time data '2025-07-22 00:00:00+00:00' does not
match format '%Y-%m-%d %H:%M:%S.%f'